feat: create Help component to streamline HelpPage functionality
